Make the CloudWatch config a singleton
Makes the CloudWatch config a singleton object. This prevents the
CloudWatch session from being re-initialised every time we check an
alert.
Creating a new CloudWatchContext threw away all the profiles we
previously created and meant to reuse. This is suspected to cause issues
with the AWS metadata API, which we suspect rate limits us  when querying
a lot of CloudWatch metrics.